+++ lib/oneview-sdk/resource/api300/server_profile.rb
@@ -9,7 +9,6 @@
 # CONDITIONS OF ANY KIND, either express or implied. See the License for the specific
 # language governing permissions and limitations under the License.
 
-require_relative 'resource'
 require_relative '../api200'
 
 module OneviewSDK